@@ -19,21 +19,24 @@
         }
     ],
     "require": {
-        "php": ">=7.0.0",
+        "php": "^8.0.2",
         "ext-openssl": "*",
         "ext-json": "*",
         "ext-mbstring": "*",
         "ext-iconv": "*",
+        "ext-libxml": "*",
+        "ext-zip": "*",
         "ext-fileinfo": "*",
-        "nesbot/carbon": ">=1.0",
+        "nesbot/carbon": "^2.62.1",
         "symfony/http-foundation": ">=2.8.0",
         "illuminate/pagination": ">=5.0.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "~4.0"
+        "phpunit/phpunit": "^9.5.10"
     },
     "suggest": {
-        "symfony/mime": "Recomended for better extension support"
+        "symfony/mime": "Recomended for better extension support",
+        "symfony/var-dumper": "Usefull tool for debugging"
     },
     "autoload": {
         "psr-4": {
