@@ -14,9 +14,9 @@
         }
     ],
     "require": {
-        "php": "^7.3|^8.0",
-        "psr/container": "^1.0",
-        "psr/simple-cache": "^1.0"
+        "php": "^8.1",
+        "psr/container": "^1.1.1|^2.0.1",
+        "psr/simple-cache": "^1.0|^2.0|^3.0"
     },
     "autoload": {
         "psr-4": {
@@ -25,7 +25,7 @@
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "8.x-dev"
+            "dev-master": "10.x-dev"
         }
     },
     "config": {
