@@ -14,11 +14,11 @@
         }
     ],
     "require": {
-        "php": "^7.3|^8.0",
-        "ext-json": "*",
-        "illuminate/collections": "^8.0",
-        "illuminate/contracts": "^8.0",
-        "illuminate/support": "^8.0"
+        "php": "^8.1",
+        "ext-filter": "*",
+        "illuminate/collections": "^10.0",
+        "illuminate/contracts": "^10.0",
+        "illuminate/support": "^10.0"
     },
     "autoload": {
         "psr-4": {
@@ -27,7 +27,7 @@
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "8.x-dev"
+            "dev-master": "10.x-dev"
         }
     },
     "config": {
