@@ -14,15 +14,17 @@
         }
     ],
     "require": {
-        "php": "^7.3|^8.0",
-        "ext-json": "*",
+        "php": "^8.1",
+        "ext-ctype": "*",
+        "ext-filter": "*",
         "ext-mbstring": "*",
-        "doctrine/inflector": "^1.4|^2.0",
-        "illuminate/collections": "^8.0",
-        "illuminate/contracts": "^8.0",
-        "illuminate/macroable": "^8.0",
-        "nesbot/carbon": "^2.53.1",
-        "voku/portable-ascii": "^1.6.1"
+        "doctrine/inflector": "^2.0",
+        "illuminate/collections": "^10.0",
+        "illuminate/conditionable": "^10.0",
+        "illuminate/contracts": "^10.0",
+        "illuminate/macroable": "^10.0",
+        "nesbot/carbon": "^2.67",
+        "voku/portable-ascii": "^2.0"
     },
     "conflict": {
         "tightenco/collect": "<5.5.33"
@@ -37,15 +39,16 @@
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "8.x-dev"
+            "dev-master": "10.x-dev"
         }
     },
     "suggest": {
-        "illuminate/filesystem": "Required to use the composer class (^8.0).",
-        "league/commonmark": "Required to use Str::markdown() and Stringable::markdown() (^1.3|^2.0.2).",
-        "ramsey/uuid": "Required to use Str::uuid() (^4.2.2).",
-        "symfony/process": "Required to use the composer class (^5.4).",
-        "symfony/var-dumper": "Required to use the dd function (^5.4).",
+        "illuminate/filesystem": "Required to use the composer class (^10.0).",
+        "league/commonmark": "Required to use Str::markdown() and Stringable::markdown() (^2.0.2).",
+        "ramsey/uuid": "Required to use Str::uuid() (^4.7).",
+        "symfony/process": "Required to use the composer class (^6.2).",
+        "symfony/uid": "Required to use Str::ulid() (^6.2).",
+        "symfony/var-dumper": "Required to use the dd function (^6.2).",
         "vlucas/phpdotenv": "Required to use the Env class and env helper (^5.4.1)."
     },
     "config": {
